+++ clicache/src/CacheableObjectArray.hpp
@@ -119,9 +119,9 @@ namespace Apache
         /// <summary>
         /// return the size of this object in bytes
         /// </summary>
-        virtual property System::UInt32 ObjectSize
+        virtual property System::UInt64 ObjectSize
         {
-          virtual System::UInt32 get();
+          virtual System::UInt64 get();
         }
 
         /// <summary>
